解决了一下代码里的警告 最终没有采用那种剪枝更好的求交点的算法 Signed-off-by: manageryzy <manageryzy@live.cn>

+++ lineCycle/draw/drawGDI.cpp
@@ -113,7 +113,7 @@ namespace drawGDI
 		for (unsigned int i = 0; i < size; i++)
 		{
 			CArc * l = cutArcList.at(i);
-			Arc(hdc, (int)(l->x - l->r), (int)(l->y - l->r), (int)(l->x + l->r), (int)(l->y + l->r), _cosTable[((int)(l->end * 10000)) % 62852] * l->r + l->x, _sinTable[((int)(l->end * 10000)) % 62852] * l->r + l->y, _cosTable[((int)(l->begin * 10000)) % 62852] * l->r + l->x, _sinTable[((int)(l->begin * 10000)) % 62852] * l->r + l->y);
+			Arc(hdc, (int)(l->x - l->r), (int)(l->y - l->r), (int)(l->x + l->r), (int)(l->y + l->r), (int)(_cosTable[((int)(l->end * 10000)) % 62852] * l->r + l->x), (int)(_sinTable[((int)(l->end * 10000)) % 62852] * l->r + l->y), (int)(_cosTable[((int)(l->begin * 10000)) % 62852] * l->r + l->x), (int)(_sinTable[((int)(l->begin * 10000)) % 62852] * l->r + l->y));
 		}
 
 		logMsg(L"圆绘制结束，开始绘制边界多边形");
